Naming updated and integration into MAF
